既にあるなら上書きしない

+++ bin/deploy.sh
@@ -7,10 +7,12 @@ function create_symlink() {
   local -r root=$(dirname $(realpath $0) | sed "s/\(${repo_name}\).*/\1/")
 
   ls -A ${root}/home | while read -r file; do
+    [[ -f "${HOME}/home/${file}" ]] && continue
     ln -s "${root}/home/${file}" "${HOME}/"
   done
 
   ls -A ${root}/config | while read -r dir; do
+    [[ -d "${HOME}/.config/${dir}" ]] && continue
     ln -s "${root}/config/${dir}" "${HOME}/.config/"
   done
 }
